INTERLOK-3714 #comment Abstract common service behaviour to CassandraServiceImp.
Supports shared connection and start/stop the connection when the service start/stop.
Add some annotations for the UI.
