+++ src/main/java/com/adaptris/core/cassandra/CassandraQueryService.java
@@ -3,15 +3,13 @@ package com.adaptris.core.cassandra;
 import javax.validation.Valid;
 import javax.validation.constraints.NotNull;
 
+import com.adaptris.annotation.AdapterComponent;
 import com.adaptris.annotation.AutoPopulated;
-import com.adaptris.core.AdaptrisConnection;
+import com.adaptris.annotation.ComponentProfile;
+import com.adaptris.annotation.DisplayOrder;
 import com.adaptris.core.AdaptrisMessage;
 import com.adaptris.core.ConnectedService;
-import com.adaptris.core.CoreException;
-import com.adaptris.core.ServiceException;
-import com.adaptris.core.ServiceImp;
 import com.adaptris.core.cassandra.params.CachedStatementPrimer;
-import com.adaptris.core.cassandra.params.CassandraParameterApplicator;
 import com.adaptris.core.cassandra.params.NullParameterApplicator;
 import com.adaptris.core.cassandra.params.NullStatementPrimer;
 import com.adaptris.core.cassandra.params.StatementPrimer;
@@ -54,86 +52,36 @@ import com.thoughtworks.xstream.annotations.XStreamAlias;
  * Finally the results of the query can be stored in the {@link AdaptrisMessage}, the format and location of which can be configured
  * using {@link ResultSetTranslator}.
  * </p>
- * 
+ *
  * @author amcgrath
  * @config cassandra-query-service
- * @license ENTERPRISE
  */
 @XStreamAlias("cassandra-query-service")
-public class CassandraQueryService extends ServiceImp implements ConnectedService {
+@AdapterComponent
+@ComponentProfile(summary = "Execute CQL to query data from the databases tables", recommended = {
+    CassandraConnection.class }, tag = "cassandra")
+@DisplayOrder(order = { "connection", "statement" })
+public class CassandraQueryService extends CassandraServiceImp implements ConnectedService {
 
-  @NotNull
-  @Valid
-  private AdaptrisConnection connection;
   @NotNull
   @AutoPopulated
   @Valid
   private ResultSetTranslator resultSetTranslator;
-  @NotNull
-  @Valid
-  private DataInputParameter<String> statement;
-  @NotNull
-  @Valid
-  @AutoPopulated
-  private CassandraParameterApplicator parameterApplicator;
-  @NotNull
-  @Valid
-  @AutoPopulated
-  private StatementParameterList parameterList;
-  @NotNull
-  @Valid
-  @AutoPopulated
-  private StatementPrimer statementPrimer;
-    
+
   public CassandraQueryService() {
-    this.setParameterApplicator(new NullParameterApplicator());
-    this.setParameterList(new StatementParameterList());
-    this.setResultSetTranslator(new XmlPayloadTranslator());
-    this.setStatementPrimer(new NullStatementPrimer());
-  }
-  
-  @Override
-  public void doService(AdaptrisMessage message) throws ServiceException {
-    try {          
-      Session session = (((CassandraConnection) connection).getSession());
-      BoundStatement boundStatement = this.getParameterApplicator().applyParameters(session, message, this.getParameterList(), this.getStatement().extract(message));
-      ResultSet results = session.execute(boundStatement);
-      
-      JdbcResult result = new ResultBuilder().setHasResultSet(true).setResultSet(results).build();
-      resultSetTranslator.translate(result, message);
-    } catch(Exception ex) {
-      throw new ServiceException(ex);
-    }
+    setParameterApplicator(new NullParameterApplicator());
+    setParameterList(new StatementParameterList());
+    setResultSetTranslator(new XmlPayloadTranslator());
+    setStatementPrimer(new NullStatementPrimer());
   }
 
   @Override
-  public void prepare() throws CoreException {
-  }
+  public void doCassandraService(Session session, AdaptrisMessage message) throws Exception {
+    BoundStatement boundStatement = getParameterApplicator().applyParameters(session, message, getParameterList(), getStatement().extract(message));
+    ResultSet results = session.execute(boundStatement);
 
-
-  @Override
-  protected void initService() throws CoreException {
-    this.getParameterApplicator().setStatementPrimer(this.getStatementPrimer());
-  }
-
-  @Override
-  protected void closeService() {
-  }
-
-  @Override
-  public void start() throws CoreException {
-  }
-  
-  @Override
-  public void stop() {
-  }
-  
-  public AdaptrisConnection getConnection() {
-    return connection;
-  }
-
-  public void setConnection(AdaptrisConnection connection) {
-    this.connection = connection;
+    JdbcResult result = new ResultBuilder().setHasResultSet(true).setResultSet(results).build();
+    resultSetTranslator.translate(result, message);
   }
 
   public ResultSetTranslator getResultSetTranslator() {
@@ -142,39 +90,6 @@ public class CassandraQueryService extends ServiceImp implements ConnectedServic
 
   public void setResultSetTranslator(ResultSetTranslator resultSetTranslator) {
     this.resultSetTranslator = resultSetTranslator;
-  }
-
-  public DataInputParameter<String> getStatement() {
-    return statement;
-  }
-
-  public void setStatement(DataInputParameter<String> statement) {
-    this.statement = statement;
-  }
-
-  public CassandraParameterApplicator getParameterApplicator() {
-    return parameterApplicator;
-  }
-
-  public void setParameterApplicator(
-      CassandraParameterApplicator parameterApplicator) {
-    this.parameterApplicator = parameterApplicator;
-  }
-
-  public StatementParameterList getParameterList() {
-    return parameterList;
-  }
-
-  public void setParameterList(StatementParameterList parameterList) {
-    this.parameterList = parameterList;
-  }
-
-  public StatementPrimer getStatementPrimer() {
-    return statementPrimer;
-  }
-
-  public void setStatementPrimer(StatementPrimer statementPrimer) {
-    this.statementPrimer = statementPrimer;
   }
 
 }
